[graph] Move missing box detection to layout from useAutoBox
This solves a few conceptual issues with regards to whether useAutoBox should query for missing boxes on first render only, or simply whenever a request is made (the latter wins).

We also do this via the (all important) BoxContext - same mechanism as for later dynamic boxes.

The layout anyhow iterates on connections and the situation of an end not having a box was not handled.

The danger here is - what if we are not using the connection layout? Would other layouts need to replicate this mechanism?