[graph] Pass ref from withAutoBox to GraphBase
This is in order to have a reference to the graph element (so we can have similar ids under different graphs, and so relative positioning is more efficient.

There is currently no need to support ref in GraphBase other than that in withAutoBox. So we use forwardedRef prop rather than React.forwardRef. This greatly simplifies the code.

Also, per Abramov's comment (https://github.com/facebook/react/issues/15306#issuecomment-479544799) refs should be back in props soon.