[graph] No more scaling, not more html within foreignObject
Scaling of the whole graph was done by providing viewbox, setting preserveAspectRatio, and putting HTML nodes within a foreignObject.

There were a host of issues with this approach.

First, UX, scalled graphs mean smaller hit targets.

Then:

There was misalignments when scaled because the client coordinates where based on the screen (not scaled)  coordinates. Similarly, dragging would odd results (again because the mouse coordinates are not transformed to account for scale).

Then, both on iPhone (chrome) and desktop (Safari), dragging element within foreignObject leave 'trails' - the canvas does not seem to update the edges of the moved region.