[connections] Removing caching of shape in connections
We were caching the shape in the connection from resolveAnchors so the connection wouldn't have to re-claculate it.

But onces resolved the shape would remain on the store, and the trimming of the arrow head would mutate that shape time and again on the server upon page refresh.

This is because the store is ultimately created once when editor is created - which is when the module it is in is required. So that store data remains between server-side renders.